common: cache total RAM as `uint64_t`
Detect total RAM once so automatic cache selection and the oversized-cache warning use the same value.
Database cache arithmetic now uses `uint64_t`, so return the byte count directly and remove the 32-bit `size_t` clamp.

+++ src/common/system_ram.h
@@ -5,12 +5,12 @@
 #ifndef BITCOIN_COMMON_SYSTEM_RAM_H
 #define BITCOIN_COMMON_SYSTEM_RAM_H
 
-#include <cstddef>
+#include <cstdint>
 #include <optional>
 
 /**
  * Return the total RAM available on the current system, if detectable.
  */
-std::optional<size_t> TryGetTotalRam();
+std::optional<uint64_t> TryGetTotalRam();
 
 #endif // BITCOIN_COMMON_SYSTEM_RAM_H
